Move interfaces/* -> source-decorators/*
